Split services into dedicated file for System Bridge
